Support for the latest version of the API (2.3.0.1)
Literally most this does is comment everything that uses the old NameMenu from the API before 2.0.

You can no longer change options here because the NameMenu was removed. I should have made it so it uses instead the new Option Menu API but i'm not in the mood to debug errors now lol.